pyproject.toml
[project]
name = "kaseb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiofiles>=24.1.0",
  "asyncio>=4.0.0",
  "fastapi>=0.119.0",
  "httpx>=0.28.1",
  "ipykernel>=6.30.1",
  "matplotlib>=3.10.7",
  "mlflow>=3.5.0",
  "motor>=3.7.1",
  "pandas>=2.3.2",
  "pwdlib[argon2]>=0.3.0",
  "pyjwt>=2.10.1",
  "pymongo>=4.15.1",
  "pytest>=9.0.1",
  "pytest-asyncio>=1.3.0",
  "python-decouple>=3.8",
  "python-multipart>=0.0.20",
  "scikit-learn>=1.7.2",
  "torch>=2.7.1",
  "transformers>=4.56.2",
  "uvicorn>=0.38.0",
]
[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
